moved all player related files into players package
